Parse fabric loader version from modloader ID
CurseForge doesn't (or no longer) users the yarn property in the
manifest to specify the fabric version. Instead they now include the
version in id property, ex: `fabric-0.11.6`.

When importing an instance we'll check for the yarn property first, for
backwards compatibility, but will then try to extract it from the id if
it's missing.

For exporting we'll now include the loader version in the id and omit
the yarn property, exactly like CurseForge.

Fixes #942