Implement ratelimiting for requests
This PR implements ratelimiting for requests generated by GDLauncher, as
to prevent spamming the API which likely causes the errors while
downloading that happen very often (at least in my experience).

This is probably not the best way to do this, but I've found it's quite
effective so far. This implementation creates an axios shim (to have a
global ratelimit, not just per-request type) in
`src/common/axios-shim.js` that just wraps an `axios.create()` call in a
`rateLimit` (from `axios-rate-limit`).

This currently sets the maximum requests/sec to 50. This can likely be
changed, but should not be put too high as it may cause the same
problems that already existed. Frankly, it's better to download a bit
slower than you could otherwise, than causing problems and not being
able to download at all.

Some of these problems are indeed due to CurseForge's API, but many of
them likely stem (though I'm not entirely sure, my experiences before
and after this change seem to prove this) from a high frequency of
requests to the CurseForge and GDLauncher API.

Ideally, this rate-limiting would be implemented on the API side and be
a bit more robust than this, but that's not always possible.

This is not a solution to all the API problems, but it is an important
one. It's important that this be implemented somehow on the API side, to
prevent bad actors from intentionally slowing down the API. However, for
most users they will use the official GDLauncher build which will
contain this fix.

We can't change the CurseForge API, perhaps, but we can change the
GDLauncher API (though not us contributors directly, as I don't believe
the source is public), and we can change how we handle problems and
requests.

I'm not sure what exactly our strategy for retrying requests is, but
from my experience it seems like it hasn't worked very well. This may
need to change, it may not. It is possible I may have a misunderstanding
of the code and what has occurred.